Change compatible IDEA versions
patchPluginXml tasks patches plugin.xml file and sometimes does it not
the way which is intuitively expected. E.g. if sinceBuild is set and
untilBuild is not, then it takes value for untilBuild from
`intellij.version` prop.

The change hardly specifies since version in plugin.xml, and leaves
until version empty.